requirements.txt
langgraph
langchain-groq
langchain-core
python-dotenv
deep-translator
streamlit